Added utility library for reading and writing packets.
This commit adds the W2HP Packet Translator library to the project,
which is essential for the development of the Work Hard Play Hard
Protocol. All of annoying to deal with implementation details for
writing and reading packets as byte arrays are encapsulated into easy to
use classes by this library.

This commit also includes an initial protocol proposal for packets that
may be sent between the client and the server as a .JSON file named
protocol-dictionary.json . The W2HP Packet Translator library has the
ability to parse files that are formatted in the same way as the
protocol-dictionary.json file in order to ensure that bindings between
enumerator constants and fields are accurate to provided specification.