Converted W2HP Packet Translator to Maven project and updated getter
Converted W2HP Packet Translator to Maven project

I've started work on the server side of the application and found that
using Maven can help significantly with managing dependencies, so I'm
converting the W2HP Packet Translator to a Maven project for future
project management.

Updated getter

The previous method for getting PacketInterface objects required an
instance of the Packet enumerator, which may have led to confusing code
as a Packet constant would have had to been selected when the class was
already specific enough. Using Class<P> as the type for the bindings
argument clarifies the purpose of the argument in code.